feat(server): improve application prompt and dependabot auto merge workflow

+++ packages/schemas/src/application/match.schemas.ts
@@ -8,7 +8,7 @@ export const SeniorityFitSchema = z.enum(SENIORITY_FIT_VALUES);
 export type SeniorityFit = z.infer<typeof SeniorityFitSchema>;
 
 export const MatchSchema = z.object({
-  score: z.number().min(0).max(100),
+  score: z.number().int().min(0).max(100),
   confidence: ConfidenceSchema,
   strengths: z.array(z.string()),
   missing_skills: z.array(z.string()),
